feat: rework logic to use aik and new dev toolbar stuff
