+++ package/env.d.ts
@@ -1,6 +1,6 @@
 /// <reference types="astro/client" />
 
-declare module "virtual:@astrojs/tailwind/dev-overlay" {
+declare module "virtual:astro-tailwind-config-viewer/internal" {
 	export const viewerLink: string;
 	export const overlayMode: Exclude<
 		import("./astro-tailwind-config-viewer.js").Options["viewer"],
